move store to folder
